backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-storages
boto3
celery
redis
stripe
requests
Pillow
python-decouple
django-filter
drf-yasg
python-dateutil
django-admin-interface
django-colorfield
setuptools<81
wheel
google-auth
google-auth-oauthlib
google-auth-httplib2
facebook-sdk

